+++ nf_core/pipelines/schema.py
@@ -327,11 +327,32 @@ class PipelineSchema:
             schema_no_required = copy.deepcopy(self.schema)
             if "required" in schema_no_required:
                 schema_no_required.pop("required")
+            for keyword in ["allOf", "anyOf", "oneOf"]:
+                if keyword in schema_no_required:
+                    for i, kw_content in enumerate(schema_no_required[keyword]):
+                        if "required" in kw_content:
+                            schema_no_required[keyword][i].pop("required")
+                    schema_no_required[keyword] = [
+                        kw_content for kw_content in schema_no_required[keyword] if kw_content
+                    ]
+                    if not schema_no_required[keyword]:
+                        schema_no_required.pop(keyword)
             for group_key, group in schema_no_required.get(self.defs_notation, {}).items():
                 if "required" in group:
                     schema_no_required[self.defs_notation][group_key].pop("required")
+                for keyword in ["allOf", "anyOf", "oneOf"]:
+                    if keyword in group:
+                        for i, kw_content in enumerate(group[keyword]):
+                            if "required" in kw_content:
+                                schema_no_required[self.defs_notation][group_key][keyword][i].pop("required")
+                        schema_no_required[self.defs_notation][group_key][keyword] = [
+                            kw_content for kw_content in group[keyword] if kw_content
+                        ]
+                        if not group[keyword]:
+                            schema_no_required[self.defs_notation][group_key].pop(keyword)
             jsonschema.validate(self.schema_defaults, schema_no_required)
         except jsonschema.exceptions.ValidationError as e:
+            log.debug(f"Complete error message:\n{e}")
             raise AssertionError(f"Default parameters are invalid: {e.message}")
         for param, default in self.schema_defaults.items():
             if default in ("null", "", None, "None") or default is False: